chore: Move version types to own file
